+++ apps/commerce-api/src/test/java/com/loopers/interfaces/api/user/UserInfoE2ETest.java
@@ -1,12 +1,9 @@
-package com.loopers.domain.example;
+package com.loopers.interfaces.api.user;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.loopers.application.example.user.UserFacade;
 import com.loopers.application.example.user.UserResponse;
 import com.loopers.interfaces.api.ApiResponse;
-import com.loopers.interfaces.api.user.LoginRequest;
-import com.loopers.interfaces.api.user.RegisterUserRequest;
 import com.loopers.utils.DatabaseCleanUp;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -17,12 +14,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class UserLoginE2ETest {
+class UserInfoE2ETest {
 
     @Autowired
     private TestRestTemplate restTemplate;
@@ -31,8 +29,7 @@ class UserLoginE2ETest {
     private DatabaseCleanUp databaseCleanUp;
 
     private static final String REGISTER_URL = "/api/v1/users";
-    private static final String LOGIN_URL = "/api/v1/users/login";
-    private UserFacade userFacade;
+    private static final String USER_INFO_URL = "/api/v1/users/me";
 
     @AfterEach
     void tearDown() {
@@ -40,57 +37,55 @@ class UserLoginE2ETest {
     }
 
     @Nested
-    class Login {
+    class FindUserInfo {
 
         @Test
-        @DisplayName("존재하는 유저 ID로 로그인하면 X-USER-ID 헤더를 반환한다.")
-        void login_success() {
+        @DisplayName("존재하는 유저의 정보를 조회하면 200 OK와 유저 정보를 반환한다.")
+        void findUserInfo_success() {
             // given
-
-
-            RegisterUserRequest registerRequest = new RegisterUserRequest(
+            RegisterUserRequest request = new RegisterUserRequest(
                 "kth4909", "kth@loopers.com", "M", "1999-10-23");
 
-            restTemplate.postForEntity(REGISTER_URL, registerRequest, Void.class);
+            restTemplate.postForEntity(REGISTER_URL, request, Void.class);
 
-            LoginRequest loginRequest = new LoginRequest("kth4909");
+            HttpHeaders headers = new HttpHeaders();
+            headers.add("X-USER-ID", "kth4909");
+            HttpEntity<Void> httpRequest = new HttpEntity<>(headers);
 
             // when
             ResponseEntity<ApiResponse<UserResponse>> response = restTemplate.exchange(
-                LOGIN_URL,
-                HttpMethod.POST,
-                new HttpEntity<>(loginRequest),
-                new ParameterizedTypeReference<>() {
-                }
+                USER_INFO_URL,
+                HttpMethod.GET,
+                httpRequest,
+                new ParameterizedTypeReference<>() {}
             );
 
             // then
             assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-            assertThat(response.getHeaders().getFirst("X-USER-ID")).isEqualTo("kth4909");
             assertThat(response.getBody()).isNotNull();
             assertThat(response.getBody().data().userId()).isEqualTo("kth4909");
         }
 
         @Test
-        @DisplayName("존재하지 않는 유저 ID로 로그인하면 404를 반환한다.")
-        void login_fail_userNotFound() {
+        @DisplayName("존재하지 않는 유저 ID로 조회하면 404 Not Found를 반환한다.")
+        void findUserInfo_fail_notFound() {
             // given
-            LoginRequest loginRequest = new LoginRequest("unknownUser");
+            HttpHeaders headers = new HttpHeaders();
+            headers.add("X-USER-ID", "unknownUser");
+            HttpEntity<Void> httpRequest = new HttpEntity<>(headers);
 
             // when
             ResponseEntity<ApiResponse<Object>> response = restTemplate.exchange(
-                LOGIN_URL,
-                HttpMethod.POST,
-                new HttpEntity<>(loginRequest),
-                new ParameterizedTypeReference<>() {
-                }
+                USER_INFO_URL,
+                HttpMethod.GET,
+                httpRequest,
+                new ParameterizedTypeReference<>() {}
             );
 
             // then
             assertThat(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
             assertThat(response.getBody()).isNotNull();
             assertThat(response.getBody().meta().result()).isEqualTo(ApiResponse.Metadata.Result.FAIL);
-            assertThat(response.getBody().data()).isNull();
         }
     }
 }
